move SlotState into its own file
